+++ src/main/java/de/maxhenkel/voicechat/VoicechatClient.java
@@ -3,14 +3,20 @@ package de.maxhenkel.voicechat;
 import de.maxhenkel.voicechat.config.ClientConfig;
 import de.maxhenkel.voicechat.config.ConfigBuilder;
 import de.maxhenkel.voicechat.config.PlayerVolumeConfig;
+import de.maxhenkel.voicechat.net.Packets;
 import de.maxhenkel.voicechat.voice.client.ClientVoiceEvents;
+import io.netty.buffer.Unpooled;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
+import net.fabricmc.fabric.api.client.networking.v1.ClientLoginNetworking;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.options.KeyBinding;
+import net.minecraft.network.PacketByteBuf;
 import org.lwjgl.glfw.GLFW;
+
+import java.util.concurrent.CompletableFuture;
 
 @Environment(EnvType.CLIENT)
 public class VoicechatClient implements ClientModInitializer {
@@ -27,6 +33,18 @@ public class VoicechatClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
+        ClientLoginNetworking.registerGlobalReceiver(Packets.INIT, (client, handler, buf, listenerAdder) -> {
+            int serverCompatibilityVersion = buf.readInt();
+
+            if (serverCompatibilityVersion != Voicechat.COMPATIBILITY_VERSION) {
+                Voicechat.LOGGER.warn("Incompatible voice chat version (server={}, client={})", serverCompatibilityVersion, Voicechat.COMPATIBILITY_VERSION);
+            }
+
+            PacketByteBuf buffer = new PacketByteBuf(Unpooled.buffer());
+            buffer.writeInt(Voicechat.COMPATIBILITY_VERSION);
+            return CompletableFuture.completedFuture(buffer);
+        });
+
         ConfigBuilder.create(MinecraftClient.getInstance().runDirectory.toPath().resolve("config").resolve(Voicechat.MODID).resolve("voicechat-client.properties"), builder -> CLIENT_CONFIG = new ClientConfig(builder));
         VOLUME_CONFIG = new PlayerVolumeConfig(MinecraftClient.getInstance().runDirectory.toPath().resolve("config").resolve("voicechat-volumes.properties"));