+++ src/main/java/de/maxhenkel/voicechat/net/InitPacket.java
@@ -1,5 +1,6 @@
 package de.maxhenkel.voicechat.net;
 
+import de.maxhenkel.voicechat.config.ServerConfig;
 import net.minecraft.network.PacketByteBuf;
 
 import java.util.UUID;
@@ -8,16 +9,16 @@ public class InitPacket {
 
     private UUID secret;
     private int serverPort;
-    private int sampleRate;
+    private ServerConfig.Codec codec;
     private int mtuSize;
     private double voiceChatDistance;
     private double voiceChatFadeDistance;
     private int keepAlive;
 
-    public InitPacket(UUID secret, int serverPort, int sampleRate, int mtuSize, double voiceChatDistance, double voiceChatFadeDistance, int keepAlive) {
+    public InitPacket(UUID secret, int serverPort, ServerConfig.Codec codec, int mtuSize, double voiceChatDistance, double voiceChatFadeDistance, int keepAlive) {
         this.secret = secret;
         this.serverPort = serverPort;
-        this.sampleRate = sampleRate;
+        this.codec = codec;
         this.mtuSize = mtuSize;
         this.voiceChatDistance = voiceChatDistance;
         this.voiceChatFadeDistance = voiceChatFadeDistance;
@@ -32,8 +33,8 @@ public class InitPacket {
         return serverPort;
     }
 
-    public int getSampleRate() {
-        return sampleRate;
+    public ServerConfig.Codec getCodec() {
+        return codec;
     }
 
     public int getMtuSize() {
@@ -53,13 +54,13 @@ public class InitPacket {
     }
 
     public static InitPacket fromBytes(PacketByteBuf buf) {
-        return new InitPacket(buf.readUuid(), buf.readInt(), buf.readInt(), buf.readInt(), buf.readDouble(), buf.readDouble(), buf.readInt());
+        return new InitPacket(buf.readUuid(), buf.readInt(), ServerConfig.Codec.values()[buf.readByte()], buf.readInt(), buf.readDouble(), buf.readDouble(), buf.readInt());
     }
 
     public void toBytes(PacketByteBuf buf) {
         buf.writeUuid(secret);
         buf.writeInt(serverPort);
-        buf.writeInt(sampleRate);
+        buf.writeByte(codec.ordinal());
         buf.writeInt(mtuSize);
         buf.writeDouble(voiceChatDistance);
         buf.writeDouble(voiceChatFadeDistance);