+++ src/main/java/de/maxhenkel/voicechat/voice/common/MicPacket.java
@@ -5,9 +5,11 @@ import net.minecraft.network.PacketByteBuf;
 public class MicPacket implements Packet<MicPacket> {
 
     private byte[] data;
+    private long sequenceNumber;
 
-    public MicPacket(byte[] data) {
+    public MicPacket(byte[] data, long sequenceNumber) {
         this.data = data;
+        this.sequenceNumber = sequenceNumber;
     }
 
     public MicPacket() {
@@ -18,15 +20,21 @@ public class MicPacket implements Packet<MicPacket> {
         return data;
     }
 
+    public long getSequenceNumber() {
+        return sequenceNumber;
+    }
+
     @Override
     public MicPacket fromBytes(PacketByteBuf buf) {
         MicPacket soundPacket = new MicPacket();
         soundPacket.data = buf.readByteArray();
+        soundPacket.sequenceNumber = buf.readLong();
         return soundPacket;
     }
 
     @Override
     public void toBytes(PacketByteBuf buf) {
         buf.writeByteArray(data);
+        buf.writeLong(sequenceNumber);
     }
 }