+++ src/main/java/de/maxhenkel/voicechat/voice/common/SoundPacket.java
@@ -1,6 +1,5 @@
 package de.maxhenkel.voicechat.voice.common;
 
-
 import net.minecraft.network.PacketByteBuf;
 
 import java.util.UUID;
@@ -9,10 +8,12 @@ public class SoundPacket implements Packet<SoundPacket> {
 
     private UUID sender;
     private byte[] data;
+    private long sequenceNumber;
 
-    public SoundPacket(UUID sender, byte[] data) {
+    public SoundPacket(UUID sender, byte[] data, long sequenceNumber) {
         this.sender = sender;
         this.data = data;
+        this.sequenceNumber = sequenceNumber;
     }
 
     public SoundPacket() {
@@ -27,11 +28,16 @@ public class SoundPacket implements Packet<SoundPacket> {
         return sender;
     }
 
+    public long getSequenceNumber() {
+        return sequenceNumber;
+    }
+
     @Override
     public SoundPacket fromBytes(PacketByteBuf buf) {
         SoundPacket soundPacket = new SoundPacket();
         soundPacket.sender = buf.readUuid();
         soundPacket.data = buf.readByteArray();
+        soundPacket.sequenceNumber = buf.readLong();
         return soundPacket;
     }
 
@@ -39,5 +45,6 @@ public class SoundPacket implements Packet<SoundPacket> {
     public void toBytes(PacketByteBuf buf) {
         buf.writeUuid(sender);
         buf.writeByteArray(data);
+        buf.writeLong(sequenceNumber);
     }
 }